# Conflicts:
#	decryptor/postgresql/data_encoder.go
#	decryptor/postgresql/data_encoder_test.go
#	decryptor/postgresql/pg_decryptor.go
#	decryptor/postgresql/proxy.go
#	encryptor/config/common/encryptedTypes.go
#	encryptor/config/encryptionSettings.go
#	encryptor/config/schemaStore_test.go
#	encryptor/dbDataCoder.go
#	encryptor/queryDataEncryptor.go
#	tests/test.py
#	utils/escape_format.go
#	utils/utils.go

+++ sqlparser/sql.go
@@ -653,7 +653,7 @@ const yyErrCode = 2
 const yyInitialStackSize = 16
 
 //line yacctab:1
-var yyExca = [...]int{
+var yyExca = [...]int16{
 	-1, 1,
 	1, -1,
 	-2, 0,
@@ -720,7 +720,7 @@ const yyPrivate = 57344
 
 const yyLast = 12194
 
-var yyAct = [...]int{
+var yyAct = [...]int16{
 	285, 53, 1345, 924, 679, 500, 255, 840, 1302, 858,
 	876, 549, 1176, 1148, 880, 1149, 229, 1250, 1073, 882,
 	1145, 284, 598, 918, 879, 59, 24, 548, 3, 220,
@@ -1943,7 +1943,7 @@ var yyAct = [...]int{
 	175, 139, 101, 166,
 }
 
-var yyPact = [...]int{
+var yyPact = [...]int16{
 	1451, -1000, -210, -1000, -1000, -1000, -1000, -1000, -1000, -1000,
 	-1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000,
 	-1000, -1000, -1000, -1000, 811, 831, -1000, -1000, -1000, -1000,
@@ -2084,7 +2084,7 @@ var yyPact = [...]int{
 	-201, 571, -204, 7408, -1000, 2135, 342, -1000, -1000,
 }
 
-var yyPgo = [...]int{
+var yyPgo = [...]int16{
 	0, 1056, 27, 26, 1055, 1053, 1052, 844, 842, 840,
 	1051, 1047, 1046, 1045, 1044, 1043, 1041, 1038, 1037, 1036,
 	1034, 1033, 1026, 1024, 1021, 1018, 1014, 139, 1011, 1008,
@@ -2107,7 +2107,7 @@ var yyPgo = [...]int{
 	854, 852, 851, 850, 0, 163, 848, 838, 100,
 }
 
-var yyR1 = [...]int{
+var yyR1 = [...]uint8{
 	0, 192, 193, 193, 1, 1, 1, 1, 1, 1,
 	1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
 	1, 1, 1, 1, 1, 2, 2, 2, 6, 3,
@@ -2196,7 +2196,7 @@ var yyR1 = [...]int{
 	133,
 }
 
-var yyR2 = [...]int{
+var yyR2 = [...]int8{
 	0, 2, 0, 1, 1, 1, 1, 1, 1, 1,
 	1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
 	1, 1, 1, 1, 1, 4, 6, 7, 5, 10,
@@ -2285,7 +2285,7 @@ var yyR2 = [...]int{
 	1,
 }
 
-var yyChk = [...]int{
+var yyChk = [...]int16{
 	-1000, -192, -1, -2, -6, -7, -8, -9, -10, -11,
 	-12, -13, -14, -15, -17, -18, -19, -21, -22, -23,
 	-20, -24, -25, -26, -3, -4, 6, 7, -30, 9,
@@ -2426,7 +2426,7 @@ var yyChk = [...]int{
 	272, 169, 273, -194, 274, -64, 165, -195, -195,
 }
 
-var yyDef = [...]int{
+var yyDef = [...]int16{
 	0, -2, 2, -2, 5, 6, 7, 8, 9, 10,
 	11, 12, 13, 14, 15, 16, 17, 18, 19, 20,
 	21, 22, 23, 24, 544, 0, 301, 301, 301, 301,
@@ -2567,7 +2567,7 @@ var yyDef = [...]int{
 	0, 0, 0, 0, 493, 0, 0, 226, 227,
 }
 
-var yyTok1 = [...]int{
+var yyTok1 = [...]int16{
 	1, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
@@ -2583,7 +2583,7 @@ var yyTok1 = [...]int{
 	3, 3, 3, 3, 116, 3, 128,
 }
 
-var yyTok2 = [...]int{
+var yyTok2 = [...]int16{
 	2, 3, 4, 5, 6, 7, 8, 9, 10, 11,
 	12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
 	22, 23, 24, 25, 26, 27, 28, 29, 30, 31,
@@ -2612,7 +2612,7 @@ var yyTok2 = [...]int{
 	269, 270, 271, 272, 273, 274,
 }
 
-var yyTok3 = [...]int{
+var yyTok3 = [...]uint16{
 	57600, 275, 57601, 276, 0,
 }
 
@@ -2694,9 +2694,9 @@ func yyErrorMessage(state, lookAhead int) string {
 	expected := make([]int, 0, 4)
 
 	// Look for shiftable tokens.
-	base := yyPact[state]
+	base := int(yyPact[state])
 	for tok := TOKSTART; tok-1 < len(yyToknames); tok++ {
-		if n := base + tok; n >= 0 && n < yyLast && yyChk[yyAct[n]] == tok {
+		if n := base + tok; n >= 0 && n < yyLast && int(yyChk[int(yyAct[n])]) == tok {
 			if len(expected) == cap(expected) {
 				return res
 			}
@@ -2706,13 +2706,13 @@ func yyErrorMessage(state, lookAhead int) string {
 
 	if yyDef[state] == -2 {
 		i := 0
-		for yyExca[i] != -1 || yyExca[i+1] != state {
+		for yyExca[i] != -1 || int(yyExca[i+1]) != state {
 			i += 2
 		}
 
 		// Look for tokens that we accept or reduce.
 		for i += 2; yyExca[i] >= 0; i += 2 {
-			tok := yyExca[i]
+			tok := int(yyExca[i])
 			if tok < TOKSTART || yyExca[i+1] == 0 {
 				continue
 			}
@@ -2743,30 +2743,30 @@ func yylex1(lex yyLexer, lval *yySymType) (char, token int) {
 	token = 0
 	char = lex.Lex(lval)
 	if char <= 0 {
-		token = yyTok1[0]
+		token = int(yyTok1[0])
 		goto out
 	}
 	if char < len(yyTok1) {
-		token = yyTok1[char]
+		token = int(yyTok1[char])
 		goto out
 	}
 	if char >= yyPrivate {
 		if char < yyPrivate+len(yyTok2) {
-			token = yyTok2[char-yyPrivate]
+			token = int(yyTok2[char-yyPrivate])
 			goto out
 		}
 	}
 	for i := 0; i < len(yyTok3); i += 2 {
-		token = yyTok3[i+0]
+		token = int(yyTok3[i+0])
 		if token == char {
-			token = yyTok3[i+1]
+			token = int(yyTok3[i+1])
 			goto out
 		}
 	}
 
 out:
 	if token == 0 {
-		token = yyTok2[1] /* unknown char */
+		token = int(yyTok2[1]) /* unknown char */
 	}
 	if yyDebug >= 3 {
 		__yyfmt__.Printf("lex %s(%d)\n", yyTokname(token), uint(char))
@@ -2821,7 +2821,7 @@ yystack:
 	yyS[yyp].yys = yystate
 
 yynewstate:
-	yyn = yyPact[yystate]
+	yyn = int(yyPact[yystate])
 	if yyn <= yyFlag {
 		goto yydefault /* simple state */
 	}
@@ -2832,8 +2832,8 @@ yynewstate:
 	if yyn < 0 || yyn >= yyLast {
 		goto yydefault
 	}
-	yyn = yyAct[yyn]
-	if yyChk[yyn] == yytoken { /* valid shift */
+	yyn = int(yyAct[yyn])
+	if int(yyChk[yyn]) == yytoken { /* valid shift */
 		yyrcvr.char = -1
 		yytoken = -1
 		yyVAL = yyrcvr.lval
@@ -2846,7 +2846,7 @@ yynewstate:
 
 yydefault:
 	/* default state action */
-	yyn = yyDef[yystate]
+	yyn = int(yyDef[yystate])
 	if yyn == -2 {
 		if yyrcvr.char < 0 {
 			yyrcvr.char, yytoken = yylex1(yylex, &yyrcvr.lval)
@@ -2855,18 +2855,18 @@ yydefault:
 		/* look through exception table */
 		xi := 0
 		for {
-			if yyExca[xi+0] == -1 && yyExca[xi+1] == yystate {
+			if yyExca[xi+0] == -1 && int(yyExca[xi+1]) == yystate {
 				break
 			}
 			xi += 2
 		}
 		for xi += 2; ; xi += 2 {
-			yyn = yyExca[xi+0]
+			yyn = int(yyExca[xi+0])
 			if yyn < 0 || yyn == yytoken {
 				break
 			}
 		}
-		yyn = yyExca[xi+1]
+		yyn = int(yyExca[xi+1])
 		if yyn < 0 {
 			goto ret0
 		}
@@ -2888,10 +2888,10 @@ yydefault:
 
 			/* find a state where "error" is a legal shift action */
 			for yyp >= 0 {
-				yyn = yyPact[yyS[yyp].yys] + yyErrCode
+				yyn = int(yyPact[yyS[yyp].yys]) + yyErrCode
 				if yyn >= 0 && yyn < yyLast {
-					yystate = yyAct[yyn] /* simulate a shift of "error" */
-					if yyChk[yystate] == yyErrCode {
+					yystate = int(yyAct[yyn]) /* simulate a shift of "error" */
+					if int(yyChk[yystate]) == yyErrCode {
 						goto yystack
 					}
 				}
@@ -2927,7 +2927,7 @@ yydefault:
 	yypt := yyp
 	_ = yypt // guard against "declared and not used"
 
-	yyp -= yyR2[yyn]
+	yyp -= int(yyR2[yyn])
 	// yyp is now the index of $0. Perform the default action. Iff the
 	// reduced production is ε, $1 is possibly out of range.
 	if yyp+1 >= len(yyS) {
@@ -2938,16 +2938,16 @@ yydefault:
 	yyVAL = yyS[yyp+1]
 
 	/* consult goto table to find next state */
-	yyn = yyR1[yyn]
-	yyg := yyPgo[yyn]
+	yyn = int(yyR1[yyn])
+	yyg := int(yyPgo[yyn])
 	yyj := yyg + yyS[yyp].yys + 1
 
 	if yyj >= yyLast {
-		yystate = yyAct[yyg]
+		yystate = int(yyAct[yyg])
 	} else {
-		yystate = yyAct[yyj]
-		if yyChk[yystate] != -yyn {
-			yystate = yyAct[yyg]
+		yystate = int(yyAct[yyj])
+		if int(yyChk[yystate]) != -yyn {
+			yystate = int(yyAct[yyg])
 		}
 	}
 	// dummy call; replaced with literal code
